Script Linux
After debugging it for a while, got it to work with exporting an env variable.\nTurns out that the system dbus is the only compatible library for dbus. When trying to take the vcpkg's it would make a segfault. After exporting it and calling the program from the build folder, it works correctly. Further test is needed, but this should allow me to work on the laptop on the way to school! =)